backend/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
blurhash-python==1.2.1
boto3==1.33.11
botocore==1.33.11
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.7
django-cleanup==8.0.0
django-environ==0.11.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
executing==2.0.1
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.5.0
lxml==4.9.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.8
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.12.0
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg==3.1.14
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-docx==1.1.0
python3-openid==3.2.0
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.8.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.0.7
wcwidth==0.2.12
